chore: Pre-build sanitize prom scrape_configs
