+++ tests/unit/test_config_manager.py
@@ -308,140 +308,3 @@ receivers:
     receiver_name = "prometheus/custom/otelcol/0"
     assert receiver_name in config_manager.config._config["receivers"]
     assert receiver_name in config_manager.config._config["service"]["pipelines"]["metrics/otelcol/0"]["receivers"]
-
-
-@pytest.mark.parametrize(
-    "input_replacement,expected_replacement",
-    [
-        # Single capture group ${1}
-        ("${1}.juju-73d163-3-lxd-0", "$${1}.juju-73d163-3-lxd-0"),
-        # Multiple capture groups ${1} and ${2}
-        ("${1}-${2}", "$${1}-$${2}"),
-        # Plain $var env-var-like
-        ("$MY_VAR", "$$MY_VAR"),
-        # Already escaped $$ — must be idempotent
-        ("$${1}.juju-73d163-3-lxd-0", "$${1}.juju-73d163-3-lxd-0"),
-        # No dollar signs — unchanged
-        ("static-replacement", "static-replacement"),
-        # Empty string
-        ("", ""),
-    ],
-)
-def test_sanitize_escape_prometheus_scrape_configs_parametrized(
-    input_replacement, expected_replacement
-):
-    """Test that bare $ signs in relabel_configs replacement fields are properly escaped."""
-    # GIVEN a scrape config with a relabeling rule containing a replacement value
-    scrape_configs = [
-        {
-            "job_name": "test_job",
-            "relabel_configs": [
-                {
-                    "action": "replace",
-                    "regex": "([^.]+).*",
-                    "replacement": input_replacement,
-                }
-            ],
-        }
-    ]
-
-    # WHEN sanitize_escape_prometheus_scrape_configs is called
-    result = ConfigManager.sanitize_escape_prometheus_scrape_configs(scrape_configs)
-
-    # THEN the replacement value is correctly escaped
-    assert result[0]["relabel_configs"][0]["replacement"] == expected_replacement
-
-
-def test_sanitize_escape_prometheus_scrape_configs_nested_relabel_configs():
-    """Test escaping in a realistic scrape config with relabel_configs as in the issue."""
-    # GIVEN a scrape config like those provided by ceph-mon via cos_agent
-    scrape_configs = [
-        {
-            "job_name": "juju_ceph-mon_prometheus_scrape",
-            "static_configs": [{"targets": ["10.0.0.1:9283"]}],
-            "relabel_configs": [
-                {
-                    "source_labels": ["__address__"],
-                    "action": "replace",
-                    "regex": "([^.]+).*",
-                    "replacement": "${1}.juju-73d163-3-lxd-0",
-                    "target_label": "instance",
-                }
-            ],
-        }
-    ]
-
-    # WHEN sanitize_escape_prometheus_scrape_configs is called
-    result = ConfigManager.sanitize_escape_prometheus_scrape_configs(scrape_configs)
-
-    # THEN the replacement value has its bare $ escaped
-    assert result[0]["relabel_configs"][0]["replacement"] == "$${1}.juju-73d163-3-lxd-0"
-    # AND other fields are unchanged
-    assert result[0]["job_name"] == "juju_ceph-mon_prometheus_scrape"
-    assert result[0]["static_configs"] == [{"targets": ["10.0.0.1:9283"]}]
-
-
-def test_sanitize_escape_prometheus_scrape_configs_multiple_jobs_mixed():
-    """Test that multiple jobs are handled correctly, including already-escaped ones."""
-    # GIVEN two jobs: one with bare $ signs, one already escaped
-    scrape_configs = [
-        {
-            "job_name": "job_bare",
-            "relabel_configs": [{"replacement": "${1}.host"}],
-        },
-        {
-            "job_name": "job_already_escaped",
-            "relabel_configs": [{"replacement": "$${1}.host"}],
-        },
-    ]
-
-    # WHEN sanitize_escape_prometheus_scrape_configs is called
-    result = ConfigManager.sanitize_escape_prometheus_scrape_configs(scrape_configs)
-
-    # THEN bare $ in job_bare is escaped
-    assert result[0]["relabel_configs"][0]["replacement"] == "$${1}.host"
-    # AND already-escaped $$ in job_already_escaped is unchanged
-    assert result[1]["relabel_configs"][0]["replacement"] == "$${1}.host"
-
-
-def test_sanitize_escape_prometheus_scrape_configs_no_mutation():
-    """Test that the original input is not mutated (deep copy guarantee)."""
-    # GIVEN a scrape config with bare $ signs
-    original = [{"relabel_configs": [{"replacement": "${1}.host"}]}]
-    original_copy = copy.deepcopy(original)
-
-    # WHEN sanitize_escape_prometheus_scrape_configs is called
-    ConfigManager.sanitize_escape_prometheus_scrape_configs(original)
-
-    # THEN the original input is not mutated
-    assert original == original_copy
-
-
-def test_sanitize_escape_prometheus_scrape_configs_idempotent():
-    """Test that calling the method twice returns the same result (idempotent)."""
-    # GIVEN a scrape config with bare $ signs
-    scrape_configs = [{"relabel_configs": [{"replacement": "${1}.host"}]}]
-
-    # WHEN sanitize_escape_prometheus_scrape_configs is called twice
-    once = ConfigManager.sanitize_escape_prometheus_scrape_configs(scrape_configs)
-    twice = ConfigManager.sanitize_escape_prometheus_scrape_configs(once)
-
-    # THEN the result is the same both times
-    assert once == twice
-
-
-@pytest.mark.parametrize(
-    "external_configs",
-    [
-        [{"config_yaml": "[]", "pipelines": ["metrics"]}],
-        [{"config_yaml": "receivers: []", "pipelines": ["metrics"]}],
-        ["not-a-dict"],
-    ],
-)
-def test_add_external_configs_skips_malformed_entries(external_configs):
-    config_manager = ConfigManager("otelcol/0", "otelcol", "", "", insecure_skip_verify=True)
-    initial_config = copy.deepcopy(config_manager.config._config)
-
-    config_manager.add_external_configs(external_configs)
-
-    assert config_manager.config._config == initial_config
